+++ lib/http_interceptor.dart
@@ -1,20 +1,8 @@
-library;
-
 export 'package:http/http.dart';
 
-export './extensions/base_request.dart';
-export './extensions/base_response_none.dart'
-    if (dart.library.io) './extensions/base_response_io.dart';
-export './extensions/multipart_request.dart';
-export './extensions/request.dart';
-export './extensions/response.dart';
-export './extensions/streamed_request.dart';
-export './extensions/streamed_response.dart';
-export './extensions/string.dart';
-export './extensions/uri.dart';
-export './http/http_methods.dart';
-export './http/intercepted_client.dart';
-export './http/intercepted_http.dart';
-export './models/http_interceptor_exception.dart';
-export './models/interceptor_contract.dart';
-export './models/retry_policy.dart';
+export 'src/client/intercepted_client.dart';
+export 'src/client/intercepted_http.dart';
+export 'src/extensions/string_extension.dart';
+export 'src/interceptor/http_interceptor.dart';
+export 'src/request_response/uri_extension.dart';
+export 'src/retry/retry_policy.dart';